+++ site/components/samples/inputs/FileUploadInputSample.jsx
@@ -36,7 +36,7 @@ export default class FileUploadInputSample extends ShallowComponent {
                     remote={props}
                     value={["info_test.png"]}
                     onChange={this.onChange}
-                    placeHolder="Upload Files"
+                    placeHolder="Drag and drop files here."
                     browseButtonText="Browse..."
                     multiple={true}
                 />