+++ src/progress/index.js
@@ -1,1 +1,2 @@
-export { default, default as Progress } from "./Progress";
+export {default, default as Progress} from "./Progress";
+export {default as ProgressBar} from "./ProgressBar";