+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 
 ## 1.1.15 -- Developing...
-* Showcase refresh bug fixed.
+* Showcase refresh bug fixed.0
+* BaseInput now forwards bsSize property.
 
 ## 1.1.14
 * string refs -> callback refs #29