+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * BaseInput now forwards bsSize property.
 * FileUploadInput bugs fixed
 * All components are using i18n now. #11
+* Application component added as a wrappwe component for all of the application and managing global settings (forex. language)
 
 ## 1.1.14
 * string refs -> callback refs #29